Generate genre.dat at build-time using cmake
cmake isn't the most obvious scripting language for text processing
but it does have some advantages over sed:
- It avoids another build-time dependency (e.g. on Windows)
- sed is hardly the easiest language to read anyway
- It's easy to accidentally write a sed script that works on GNU and
  not BSD (or vice-versa)

Unlike the sed generator, this does not preserve C-style comments in
the genre.dat.in file.  Since genre.dat is only used at build-time
(i.e. it's not a header shipped as part of the API) I don't think
that is important.  Otherwise it generates the same output.